Split authenticated and unauthenticated components
