+++ www/src/App.tsx
@@ -1,19 +1,17 @@
+import { useAuth0 } from '@auth0/auth0-react';
 import { Container, CssBaseline, ThemeProvider } from '@material-ui/core';
-import React, { FC } from 'react';
+import React, { FC, useEffect } from 'react';
 import { BrowserView, MobileView } from 'react-device-detect';
 import { Provider } from 'react-redux';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 
 import { Header, Section } from './components/Header';
 import { useAppSelector } from './redux/hooks';
 import adminStore from './redux/substores/admin/adminStore';
-import studentStore from './redux/substores/student/studentStore';
 import volunteerStore from './redux/substores/volunteeer/volunteerStore';
-import Community from './routes/Community';
-import Landing from './routes/Landing';
 import MobileLanding from './routes/MobileLanding';
-import MockInterview from './routes/MockInterview';
-import ResumeReview from './routes/ResumeReview';
+import StudentApp from './routes/Student';
+import UnauthenticatedApp from './routes/Unauthenticated';
 import theme from './styles/theme';
 import { COMMUNITY, COMMUNITY_ROUTE, COMPE_PLUS, MOCK_INTERVIEW, MOCK_INTERVIEW_ROUTE, RESUME_REVIEW, RESUME_REVIEW_ROUTE } from './util/constants';
 
@@ -22,27 +20,6 @@ const header_sections: Section[] = [
     { title: MOCK_INTERVIEW, url: MOCK_INTERVIEW_ROUTE },
     { title: COMMUNITY, url: COMMUNITY_ROUTE },
 ];
-
-const StudentApp: FC = () => {
-    return (
-        <Provider store={studentStore}>
-            <Switch>
-                <Route path={RESUME_REVIEW_ROUTE}>
-                    <ResumeReview />
-                </Route>
-                <Route path={MOCK_INTERVIEW_ROUTE}>
-                    <MockInterview />
-                </Route>
-                <Route path={COMMUNITY_ROUTE}>
-                    <Community />
-                </Route>
-                <Route path='/'>
-                    <Landing />
-                </Route>
-            </Switch>
-        </Provider>
-    );
-};
 
 const VolunteerApp: FC = () => {
     // TODO: Update components accordingly
@@ -77,7 +54,8 @@ const getContentByRole = (role: string) => {
 const App: FC = () => {
     const currentRole = useAppSelector((state) => state.user.currentRole);
 
-    const content = getContentByRole(currentRole);
+    const { isAuthenticated } = useAuth0();
+    const content = isAuthenticated ? getContentByRole(currentRole) : <UnauthenticatedApp />;
 
     return (
         <ThemeProvider theme={theme}>
