+++ www/src/redux/slices/userSlice.ts
@@ -1,12 +1,16 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 export const userSlice = createSlice({
     name: 'user',
     initialState: {
         roles: [],
-        currentRole: 'student',
+        currentRole: '',
     },
-    reducers: {},
+    reducers: {
+        setCurrentRole(state, action: PayloadAction<string>) {
+            state.currentRole = action.payload;
+        },
+    },
 });
 
 export default userSlice.reducer;
